z-index and overflow
This PR corrects two things:
- z-index: .bird-box wasn't displaying below .footer
It seems better to use z-index only once, on the footer, displaying it
below the rest of the content, rather than using z-index on each of the
divs which should be above. It's better also for future new blocks we
create.
- overflow: the moving .blog-posts cards make horizontal scrollbar
appear during transition, making it the web not mobile friendly in that
part.